# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Built by team C",
    "version": "1.0.0",
    "title": "Desa Sehat",
    "contact": {
      "email": "romadhonatatang@gmail.com"
    }
  },
  "host": "",
  "basePath": "/api",
...
  "paths": {
    "/user/": {
      "post": {
        "tags": [
          "User Collection"
        ],
        "summary": "create new user",
        "description": "Create a new user, the first created user is superuser",
        "operationId": "createUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/createUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "user created",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "ERROR",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "User Collection"
        ],
        "summary": "Update user profile",
        "description": "",
        "operationId": "updateProfile",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/updateProfile"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "profile updated",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "ERROR",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "get": {
        "tags": [
          "User Collection"
        ],
        "summary": "Get user data",
        "description": "",
        "operationId": "userProfile",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "ERROR",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "delete": {
        "tags": [
          "User Collection"
        ],
        "summary": "Delete current user",
        "description": "To delete the user by user's token",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "Authorization": []
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "ERROR",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/resetPassword/{token}": {
      "put": {
        "tags": [
          "User Collection"
        ],
        "summary": "change password",
        "description": "",
        "operationId": "resetPassword",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "token of user",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/resetPassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "password changed",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "ERROR",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "createUser": {
      "type": "object",
      "required": [
        "full_name",
        "email",
        "password",
        "password_confirmation",
        "gender",
        "address"
      ],
      "properties": {
        "name": {
          "type": "string",
          "example": "Tatang Romadhona"
        },
        "email": {
          "type": "string",
          "example": "romadhonatatang@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "aandeani"
        },
        "password_confirmation": {
          "type": "string",
          "example": "aandeani"
        },
        "gender": {
          "type": "string",
          "example": "Male or Female"
        },
        "address": {
          "type": "string",
          "example": "batam"
        }
      }
    },
...
    "resetPassword": {
      "type": "object",
      "required": [
        "email"
      ],
      "properties": {
        "password": {
          "type": "string",
          "example": "aandeani"
        },
        "password_confirmation": {
          "type": "string",
          "example": "aandeani"
        }
      }
    },
    "updateProfile": {
      "type": "object",
      "required": [
        "email"
      ],
      "properties": {
        "name": {
          "type": "string",
          "example": "Tatang Romadhona"
        }
      }
    },
...
  },
...
}